Disable in live courses
CCAU was never meant to be used in live courses (i.e., courses with a start date earlier than the current date) but I didn't disable it explicitly. Now I have. Thanks to Canvas for their helpful API, which made this significantly easier.

This also includes confirmation notices for irreversible changes, although that may be overkill given the above.